Updated main file solubility file, removed outdated methods. Removed outdated files.
